refactor(cli): reorder setup.py top-down + extract shared stage-line UI
Reshape src/terok/cli/commands/setup.py to match uninstall.py's
narrative order: module docstring → CLI wiring → cmd_setup → phase
helpers in execution order → systemd helpers → check-only helpers →
cmd_project_init.  Previously the file read bottom-up (palette + host
binary checks at the top, cmd_setup at the end); the reversal matches
the manifesto's "catalog matches implementation" rule so a reader
opening setup.py and uninstall.py side-by-side encounters the same
domain in the same order.

Factor the palette wrappers (_bold / _green / _red / _yellow), the
status markers (_status_label / _warn_label), and _stage_begin into a
new _setup_ui.py.  The 17-char padding in _stage_begin in particular
is a shared visual contract between setup and uninstall; SSOT pressure
applies.  setup.py previously carried the cached @cache _colour_on
variant and uninstall.py carried a fresh supports_color() per call —
the new module uses the cached variant uniformly.

Pure reorder + dedup: no behavioural change, no test modifications,
public symbols preserved (tests still import cmd_setup, dispatch, and
every _ensure_* / _disable_* helper).  All 440 CLI tests and the full
2069-test suite pass unchanged.

First of the four items from docs/brainstorming/first-run-symmetry-
followups.md (parked from the terok #786 / executor #215 / sandbox
#194 arc).
